# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AfroMarket",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7000/afro-market/v1",
      "variables": {}
    },
    {
      "url": "https://afromarket-be.herokuapp.com/afro-market/v1",
      "variables": {}
    }
  ],
  "paths": {
...
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "login user",
        "operationId": "loginuser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginuserrequest"
                  },
                  {
                    "example": {
                      "email": "babafemi.olasunmade@zummitafrica.com",
                      "password": "Wolfgang9012#"
                    }
                  }
                ]
              },
              "example": {
                "email": "babafemi.olasunmade@zummitafrica.com",
                "password": "Wolfgang9012#"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
    "/user/logout": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "log out user",
        "operationId": "logoutuser",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
    "/user/pass/forgot": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "forgot password",
        "operationId": "forgotpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/forgotpasswordrequest"
                  },
                  {
                    "example": {
                      "email": "babafemi.olasunmade@zummitafrica.com"
                    }
                  }
                ]
              },
              "example": {
                "email": "babafemi.olasunmade@zummitafrica.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
    "/user/pass/reset": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "reset password",
        "operationId": "resetpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/resetpasswordrequest"
                  },
                  {
                    "example": {
                      "newPassword": "Wolfgang9012#",
                      "otp": "179650"
                    }
                  }
                ]
              },
              "example": {
                "newPassword": "Wolfgang9012#",
                "otp": "179650"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "loginuserrequest": {
        "title": "loginuserrequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "babafemi.olasunmade@zummitafrica.com",
          "password": "Wolfgang9012#"
        }
      },
      "forgotpasswordrequest": {
        "title": "forgotpasswordrequest",
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "babafemi.olasunmade@zummitafrica.com"
        }
      },
      "resetpasswordrequest": {
        "title": "resetpasswordrequest",
        "required": [
          "newPassword",
          "otp"
        ],
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          }
        },
        "example": {
          "newPassword": "Wolfgang9012#",
          "otp": "179650"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}